+++ projects/plugins/crm/includes/ZeroBSCRM.List.php
@@ -204,6 +204,14 @@ class zeroBSCRM_list{
 			}
 		}
 
+		$screen_opts = $zbs->global_screen_options();
+
+		if ( ! empty( $screen_opts['perpage'] ) && $screen_opts['perpage'] > 0 ) {
+			$per_page = (int) $screen_opts['perpage'];
+		} else {
+			$per_page = 20;
+		}
+
         #} Refresh 2
         ?>
 
@@ -331,41 +339,28 @@ class zeroBSCRM_list{
                 <div class="ui divider"></div>
                 <?php } // if admin/can manage columns ?>
 
-                <div id="zbs-list-options-base-wrap" class="ui grid">
+								<div id="zbs-list-options-base-wrap" class="ui grid">
 
-                    <?php 
-                        # here we add stuff which is saved by screenOptions, even tho it's in its own dom elements, not sceen options area 
-                        $screenOpts = $zbs->userScreenOptions();
+									<div class="two column clearing centered row">
+										<div class="column" style="max-width:364px;">
+											<div class="ui labeled input">
+												<div class="ui label"><i class="table icon"></i>  <?php esc_html_e( 'Records per page:', 'zero-bs-crm' ); ?></div>
+												<input type="text" style="width:70px;" class="intOnly" id="zbs-screenoptions-records-per-page" value="<?php echo esc_attr( $per_page ); ?>" />
+											</div>
+										</div>
+									</div>
 
-                        // debug echo '<pre>'; print_r($screenOpts); echo '</pre>'; 
+									<div class="ui divider" style="margin-bottom:0;margin-top:0;"></div>
 
-                        // default
-                        $perPage = 20; 
-                        if (isset($screenOpts) && is_array($screenOpts) && isset($screenOpts['perpage']) && !empty($screenOpts['perpage']) && $screenOpts['perpage'] > 0) $perPage = (int)$screenOpts['perpage'];
-                    ?>
-                    <div class="two column clearing centered row">
+									<div class="two column clearing centered row">
+										<div class="column" style="max-width:364px;">
+											<button id="zbs-columnmanager-bottomsave" type="button" class="ui button black positive">
+												<i class="check square icon"></i> <?php esc_html_e( 'Save Options and Close', 'zero-bs-crm' ); ?>
+											</button>
+										</div>
+									</div>
 
-                        <div class="column" style="max-width:364px;">
-                            <div class="ui labeled input">
-										<div class="ui label"><i class="table icon"></i>  <?php esc_html_e( 'Records per page:', 'zero-bs-crm' ); ?></div>
-                                <input type="text" style="width:70px;" class="intOnly" id="zbs-screenoptions-records-per-page" value="<?php echo esc_attr( $perPage ); ?>" />
-                            </div>
-                        </div>
-
-                    </div>
-
-
-                <div class="ui divider" style="margin-bottom:0;margin-top:0;"></div>
-
-                    <div class="two column clearing centered row">
-
-                        <div class="column" style="max-width:364px;">
-									<button id="zbs-columnmanager-bottomsave" type="button" class="ui button black positive"><i class="check square icon"></i> <?php esc_html_e( 'Save Options and Close', 'zero-bs-crm' ); ?></button>
-                        </div>
-
-                    </div>
-
-                </div>
+								</div>
 
 
             </div>
@@ -401,10 +396,10 @@ class zeroBSCRM_list{
 			</div>
 			<?php
 
-			// if totals, show the wrapper
-			if ( $zbs->settings->get( 'show_totals_table' ) === 1 ) {
+			// If totals, show the wrapper. Currently only implemented in contacts
+			if ( $zbs->settings->get( 'show_totals_table' ) === 1 && $this->objType === 'customer' ) { // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
 				?>
-				<div id="jpcrm-listview-totals-box"></div>
+				<jpcrm-dashcount></jpcrm-dashcount>
 				<?php
 			}
 			##WLREMOVE
@@ -458,18 +453,6 @@ class zeroBSCRM_list{
 	            }
             }
 
-            #} Check for screen options (perpage)
-            $per_page = 20;
-            $screenOpts = $zbs->userScreenOptions();
-            if ( is_array( $screenOpts ) ) {
-
-	            if ( isset( $screenOpts['perpage'] ) ) $per_page = (int)$screenOpts['perpage'];
-	            // catch
-	            if ( $per_page < 1 ) $per_page = 20;
-
-            }
-
-
             // build options objects
             $list_view_settings = array(
 
@@ -481,7 +464,6 @@ class zeroBSCRM_list{
                 'editinline' => $allowinlineedits
 
             );
-
 
             $list_view_parameters = array(
 
@@ -502,8 +484,6 @@ class zeroBSCRM_list{
                 'pagekey' => ( isset( $zbs->pageKey ) ? esc_html( $zbs->pageKey ) : '' ),
 
             );
-
-
 
             ?>
 
@@ -573,7 +553,6 @@ class zeroBSCRM_list{
             var zbsDrawListViewAJAXBlocker = false;
             var zbsDrawListViewColUpdateBlocker = false;
             var zbsDrawListViewColUpdateAJAXBlocker = false;
-            var zbsDrawListLoadingBoxHTML = '<?php echo zeroBSCRM_UI2_loadingSegmentIncTextHTML(); ?>';
 
             var zbsObjectEmailLinkPrefix = '<?php 
 
@@ -691,28 +670,7 @@ class zeroBSCRM_list{
             var zbsClick2CallType = parseInt('<?php echo esc_url( zeroBSCRM_getSetting('clicktocalltype') ); ?>');
 
 			<?php
-			$jpcrm_listview_lang_labels = array(
-
-				'go_button'          => esc_html__( 'Go', 'zero-bs-crm' ),
-				/* translators: Placeholder is the number of selected rows. */
-				'rows_selected_x'    => esc_html__( 'Bulk actions (%s rows)', 'zero-bs-crm' ),
-				'rows_selected_1'    => esc_html__( 'Bulk actions (1 row)', 'zero-bs-crm' ),
-				'rows_selected_0'    => esc_html__( 'Bulk actions (no rows)', 'zero-bs-crm' ),
-				'zbs_edit'           => esc_html__( 'Edit', 'zero-bs-crm' ),
-				'today'              => esc_html__( 'Today', 'zero-bs-crm' ),
-				'days'               => esc_html__( 'days', 'zero-bs-crm' ),
-				'daysago'            => esc_html__( 'days ago', 'zero-bs-crm' ),
-				'notcontacted'       => esc_html__( 'Not Contacted', 'zero-bs-crm' ),
-				'yesterday'          => esc_html__( 'Yesterday', 'zero-bs-crm' ),
-				'filteredby'         => esc_html__( 'Filtered By', 'zero-bs-crm' ),
-				'notcontactedin'     => esc_html__( 'Not Contacted in', 'zero-bs-crm' ),
-				'containing'         => esc_html__( 'Containing', 'zero-bs-crm' ),
-				'couldntupdate'      => esc_html__( 'Could not update', 'zero-bs-crm' ),
-				'couldntupdatedeets' => esc_html__( 'This record could not be updated. Please try again, if this persists please let admin know.', 'zero-bs-crm' ),
-				/* translators: Placeholders are the range of the current record result and the total object count. */
-				'listview_counts'    => esc_html__( 'Showing %s of %s items', 'zero-bs-crm' ), // phpcs:ignore WordPress.WP.I18n.UnorderedPlaceholdersText
-
-			);
+			$jpcrm_listview_lang_labels = array();
 
 			// add any object-specific language labels
 			if ( count( $this->langLabels ) > 0 ) { // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
@@ -749,74 +707,54 @@ class zeroBSCRM_list{
 				$zbs_tags_for_bulk_actions = wp_json_encode( $simple_tags );
 				echo ( $zbs_tags_for_bulk_actions ? $zbs_tags_for_bulk_actions : '[]' ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
 
-				// phpcs:enable Generic.WhiteSpace.ScopeIndent.IncorrectExact,Generic.WhiteSpace.ScopeIndent.Incorrect
-										?>;
-            var zbsListViewIcos = {
+				?>;
+				var zbsListViewIcos = {};
+				// gives data used by inline editor
+				var zbsListViewInlineEdit = {
 
-                    // bulk action label icos
-                    /* has to be unicode
-                    deletetransactions: '<i class="fa fa-trash" aria-hidden="true"></i>',
-                    addtags: '<i class="fa fa-tags" aria-hidden="true"></i>',
-                    removetags: '<i class="fa fa-chain-broken" aria-hidden="true"></i>',
-                    merge: '<i class="fa fa-compress" aria-hidden="true"></i>',
-                    */
+					// for now just put contacts in here
+					customer: {
+						statuses: <?php
+							// MUST be a better way than this to get customer statuses...
+							// phpcs:disable WordPress.NamingConventions.ValidVariableName.VariableNotSnakeCase
+							global $zbsCustomerFields;
+							if ( is_array( $zbsCustomerFields['status'][3] ) ) {
+								echo wp_json_encode( $zbsCustomerFields['status'][3] );
+							} else {
+								echo '[]';
+							}
+							// phpcs:enable WordPress.NamingConventions.ValidVariableName.VariableNotSnakeCase
+						?>
+					},
 
-                    // ICONS playing up on semantic Select, so cut out for init.
-                    /* 
-                    deletetransactions: '&#xf1f8;',
-                    addtags: '&#xf1f8;',
-                    removetags: '&#xf1f8;',
-                    merge: '&#xf1f8;',
-                    */
-                    
-                    
+					owners: <?php
 
-            };
-            // gives data used by inline editor
-            var zbsListViewInlineEdit = {
+						// hardcoded customer perms atm
+						$possible_owners = zeroBS_getPossibleOwners( array( 'zerobs_admin', 'zerobs_customermgr' ), true );
+						if ( ! is_array( $possible_owners ) ) {
+								echo wp_json_encode( array() );
+						} else {
+							echo wp_json_encode( $possible_owners );
+						}
 
-                    // for now just put contacts in here
-                    customer: {
-                        statuses: <?php
+					?>
 
-                            // MUST be a better way than this to get customer statuses...
-                            global $zbsCustomerFields;
-                            if (is_array($zbsCustomerFields['status'][3])){
+					};
+					<?php
+					// Nonce for AJAX
+					echo 'var zbscrmjs_secToken = "' . esc_js( wp_create_nonce( 'zbscrmjs-ajax-nonce' ) ) . '";';
 
-                                echo json_encode($zbsCustomerFields['status'][3]);
-
-                            } else echo '[]';
-                        ?>
-                    },
-
-
-                    owners: <?php
-
-                        // hardcoded customer perms atm
-                        $zbsPossibleOwners = zeroBS_getPossibleOwners(array('zerobs_admin','zerobs_customermgr'),true);
-                        if (!is_array($zbsPossibleOwners))
-                            echo json_encode(array());
-                        else
-                            echo json_encode($zbsPossibleOwners);
-
-                    ?>
-                    
-
-            };
-            <?php #} Nonce for AJAX
-                echo "var zbscrmjs_secToken = '" . esc_js( wp_create_nonce( 'zbscrmjs-ajax-nonce' ) ) . "';";
-
-                // any last JS?
-                if (isset($this->extraJS) && !empty($this->extraJS)) echo $this->extraJS;
-            
-            ?></script>
+					// any last JS?
+					if ( isset( $this->extraJS ) && ! empty( $this->extraJS ) ) { // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
+						echo $this->extraJS; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase,WordPress.Security.EscapeOutput.OutputNotEscaped
+					}
+					// phpcs:enable Generic.WhiteSpace.ScopeIndent.IncorrectExact,Generic.WhiteSpace.ScopeIndent.Incorrect
+					?>
+					</script>
 
 					<?php
 					// phpcs:enable Squiz.PHP.EmbeddedPhp.ContentBeforeOpen
 					// phpcs:enable Squiz.PHP.EmbeddedPhp.ContentAfterEnd
-					?>
-
-        <?php
 
     } // /draw func
 
@@ -828,8 +766,8 @@ class zeroBSCRM_list{
 	public function draw_listview_header( $listview_filters ) {
 		global $zbs;
 
-		$filter_var       = 'zeroBSCRM_filterbuttons_' . $this->objType; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
-		$all_quickfilters = $GLOBALS[ $filter_var ]['all'];
+		$filter_var       = 'zeroBSCRM_filterbuttons_a' . $this->objType; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
+		$all_quickfilters = ( empty( $GLOBALS[ $filter_var ]['all'] ) ? array() : $GLOBALS[ $filter_var ]['all'] );
 		$all_tags         = $zbs->DAL->getTagsForObjType( // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
 			array(
 				'objtypeid'    => $this->objTypeID, // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
@@ -868,7 +806,7 @@ class zeroBSCRM_list{
 				}
 
 				// add tags filter if current object has tags
-				if ( count( $all_tags ) > 0 ) {
+				if ( is_array( $all_tags ) && count( $all_tags ) > 0 ) {
 					echo '<select class="filter-dropdown' . ( ! empty( $current_tag ) ? ' hidden' : '' ) . '" data-filtertype="tags">';
 					echo '<option disabled selected>' . esc_html__( 'Select tag', 'zero-bs-crm' ) . '</option>';
 					foreach ( $all_tags as $tag ) {
@@ -907,3 +845,30 @@ class zeroBSCRM_list{
 		<?php
 	}
 } // class
+
+/**
+ * Language labels for JS
+ *
+ * @param array $language_array Array of language labels.
+ */
+function jpcrm_listview_language_labels( $language_array ) { // phpcs:ignore Universal.Files.SeparateFunctionsFromOO.Mixed
+	$jpcrm_listview_lang_labels = array(
+		'click_to_sort'      => esc_html__( 'Click to sort', 'zero-bs-crm' ),
+		/* translators: Placeholder is the number of selected rows. */
+		'rows_selected_x'    => esc_html__( 'Bulk actions (%s rows)', 'zero-bs-crm' ),
+		'rows_selected_1'    => esc_html__( 'Bulk actions (1 row)', 'zero-bs-crm' ),
+		'rows_selected_0'    => esc_html__( 'Bulk actions (no rows)', 'zero-bs-crm' ),
+		'zbs_edit'           => esc_html__( 'Edit', 'zero-bs-crm' ),
+		'today'              => esc_html__( 'Today', 'zero-bs-crm' ),
+		'daysago'            => esc_html__( 'days ago', 'zero-bs-crm' ),
+		'notcontacted'       => esc_html__( 'Not Contacted', 'zero-bs-crm' ),
+		'yesterday'          => esc_html__( 'Yesterday', 'zero-bs-crm' ),
+		'couldntupdate'      => esc_html__( 'Could not update', 'zero-bs-crm' ),
+		'couldntupdatedeets' => esc_html__( 'This record could not be updated. Please try again, if this persists please let admin know.', 'zero-bs-crm' ),
+		/* translators: Placeholders are the range of the current record result and the total object count. */
+		'listview_counts'    => esc_html__( 'Showing %s of %s items', 'zero-bs-crm' ), // phpcs:ignore WordPress.WP.I18n.UnorderedPlaceholdersText
+	);
+
+	return array_merge( $language_array, $jpcrm_listview_lang_labels );
+}
+add_filter( 'zbs_globaljs_lang', 'jpcrm_listview_language_labels' );