+++ projects/packages/my-jetpack/CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.14.3] - 2023-06-12
+### Added
+- My Jetpack: populate the Jetpack AI product with feature data [#31238]
+
+## [2.14.2] - 2023-06-06
+### Changed
+- Filter out revoked licenses from the license activation options. [#31088]
+- Updated package dependencies. [#31129]
+
+## [2.14.1] - 2023-05-29
+### Added
+- My Jetpack: Add new Jetpack AI card [#30904]
+
+### Changed
+- My Jetpack: Enable Jetpack AI card [#30988]
+- My Jetpack: Update AI class to be product and not module [#30905]
+- My Jetpack: Update AI interstitial background [#30992]
+
 ## [2.14.0] - 2023-05-22
 ### Added
 - Added ability to purchase Jetpack AI monthly product [#30793]
@@ -883,6 +901,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created package
 
+[2.14.3]: https://github.com/Automattic/jetpack-my-jetpack/compare/2.14.2...2.14.3
+[2.14.2]: https://github.com/Automattic/jetpack-my-jetpack/compare/2.14.1...2.14.2
+[2.14.1]: https://github.com/Automattic/jetpack-my-jetpack/compare/2.14.0...2.14.1
 [2.14.0]: https://github.com/Automattic/jetpack-my-jetpack/compare/2.13.0...2.14.0
 [2.13.0]: https://github.com/Automattic/jetpack-my-jetpack/compare/2.12.2...2.13.0
 [2.12.2]: https://github.com/Automattic/jetpack-my-jetpack/compare/2.12.1...2.12.2